Run Wappalyzer over the browser capture HAR
Detect software and versions with Wappalyzer over the Playwright browser
HAR from webpage-capture, which sees resources loaded after JavaScript
that the static webpage-analysis HAR misses.

- Extract the shared detection logic (HAR -> Software/SoftwareInstance)
  from kat_wappalyzer into _analysis.software_from_har
- New kat_wappalyzer_capture normalizer consuming application/har+json;
  unzips the Playwright HAR and attaches SoftwareInstances to the input
  HostnameHTTPURL/IPAddressHTTPURL only, minting no OOIs for third-party
  hosts the page loaded (stays within OpenKAT scan levels)

Part of #4447.
